refactor(be): change unreadCount type from int to Long

+++ apps/api/core/src/main/java/so/morton/api/api/controller/v1/response/ChatResponse.java
@@ -10,7 +10,7 @@ public record ChatResponse(
         String title,
         List<MaskedMemberResponse> participants,
         MessageResponse lastMessage,
-        int unreadCount,
+        Long unreadCount,
         LocalDateTime createdAt,
         LocalDateTime modifiedAt
 ) {
