+++ apps/api/core/src/main/java/so/morton/api/domain/chat/ChatDetail.java
@@ -10,7 +10,7 @@ public record ChatDetail(
         String title,
         List<Member> participants,
         Message lastMessage,
-        int unreadCount,
+        Long unreadCount,
         LocalDateTime createdAt,
         LocalDateTime modifiedAt
 ) {}
